refactor: GameManager 패키지 변경
