+++ src/main/kotlin/omok/Application.kt
@@ -1,6 +1,7 @@
 package omok
 
 import omok.controller.Controller
+import omok.model.GameManager
 
 fun main() {
     val controller = Controller(GameManager())
